更改C层 S层 未完成

+++ elmboot_SpringBoot/src/main/java/com/neusoft/elmboot/controller/CartController.java
@@ -2,6 +2,10 @@ package com.neusoft.elmboot.controller;
 
 import java.util.List;
 
+import com.neusoft.elmboot.exception.BusinessException;
+import com.neusoft.elmboot.exception.ErrorCode;
+import com.neusoft.elmboot.model.vo.CartVo;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -15,23 +19,35 @@ public class CartController {
     private CartService cartService;
 
     @RequestMapping("/listCart")
-    public List<Cart> listCart(Cart cart) throws Exception {
+    public List<CartVo> listCart(Cart cart) throws Exception {
+        if (cart.getCartId() == null) {
+            throw new BusinessException(ErrorCode.SYSTEM_ERROR);
+        }
         return cartService.listCart(cart);
     }
 
     @RequestMapping("/saveCart")
     public int saveCart(Cart cart) throws Exception {
+        if (cart.getUserId() == null || cart.getBusinessId() == null ||cart.getFoodId() == null) {
+            throw new BusinessException(ErrorCode.SYSTEM_ERROR);
+        }
         return cartService.saveCart(cart);
     }
 
     @RequestMapping("/updateCart")
     public int updateCart(Cart cart) throws Exception {
+        if(){
+            throw new BusinessException(ErrorCode.SYSTEM_ERROR);
+        }
         return cartService.updateCart(cart);
     }
 
     @RequestMapping("/removeCart")
 
     public int removeCart(Cart cart) throws Exception {
+        if(){
+            throw new BusinessException(ErrorCode.SYSTEM_ERROR);
+        }
         return cartService.removeCart(cart);
     }
 }
